python/requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
